Make *ring* optional
It is a default feature.

Taking it away disables all the signature algorithms, and therefore most of the tests.
