+++ tests/integration.rs
@@ -12,6 +12,8 @@
 // ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
 // OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 
+#![cfg(feature = "ring")]
+
 use webpki::KeyUsage;
 
 static ALL_SIGALGS: &[&dyn webpki::SignatureVerificationAlgorithm] = &[
